bot/requirements.txt
python-telegram-bot==20.0
httpx==0.23.1
asyncpg==0.27.0
